requirements.txt
flask==2.0.1
Pillow==8.3.1
gunicorn==20.1.0
numpy==1.23.5
scikit-learn==1.0.2
tensorflow==2.9.3
tensorflow-hub==0.12.0
werkzeug==2.0.3
protobuf==3.20.0